+++ Assets/Scripts/UI/Menu/PlayMenu.cs
@@ -42,7 +42,7 @@ public class PlayMenu : MonoBehaviour
                 saveSlots[i].newGameImage.SetActive(false);
                 saveSlots[i].loadGameImage.SetActive(true);
                 saveSlots[i].currentAreText.text = GameManager.Instance.LoadSaveSlotData(slotIndex).playerCheckpointData.sceneName;
-                saveSlots[i].tearsOfAtonementText.text = GameManager.Instance.LoadSaveSlotData(slotIndex).playerInventoryData.tearsOfAtonement.ToString();
+                saveSlots[i].tearsOfAtonementText.text = ((int)GameManager.Instance.LoadSaveSlotData(slotIndex).playerInventoryData.tearsOfAtonement).ToString();
             }
             else
             {
